Move to newer Python, Add helpful debug information
Python 3.8 introduced the `:=` walrus operator. Previously we used
a version on alpine linux that used Python 3.6 and the version of
autopep8 that we used didn't understand walrus operator syntax.

By updating to the latest version of alpine linux we should track
the current version of Python. At the time of writing this means
Alpine Linux 3.17 and Python 3.10 (which is sufficently similar to
the current version of Python 3.11). If some major, backward
incompatible change (like the walrus operator) happens again,
it may be worth re-evaluating this policy. For example: one option
is to move to Nix/NixOS instead of docker, s.t. we can choose an
exact version of Python. It may even be worth adding that as an
option for the user of the github action.

Previously this action only reported that files changed after it
ran autopep8. Now we report a whole host of information including:
the exact versions of Python and Autopep8, the exact command run to
do the autoformat, the files that changed as a result and the diff of
those files.